meta-imx: Add optee related packages for nxp-imx8
Issue: LINUXEXEC-29533

imx-nxp-bsp is aimed at all imx platforms, so this will
bring some unnecessary packages to compile, which have no
support requirements on the imx8 platform for the time being,
so adding mx8-nxp-bsp should be more accurate for imx8
machines. In addition, the libdir is modified to nonarch_libdir
so that tee-supplicant can successfully find the plug-in after
it is started.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>