meta-imx: Restore the following files to 6.1.55 state
Issue: LINUXEXEC-30633

These files were modified during the merge process and
are now not in the 6.1.55 state. This will have an impact
on various functions. Many features have been added and
verified for version 6.1.55. In any case, we should not
modify them, otherwise the risk is unknown.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>